Refactor: move stream record reader to pixels-core/reader
